Added PaymentRequest extension methods
Added PaymentAttempts to PaymentRequestMinimal
